Extract common code into Wrapper...Symbol classes.
No behavior changes intended.
