+++ src/Compilers/CSharp/Portable/Symbols/Retargeting/RetargetingFieldSymbol.cs
@@ -19,17 +19,12 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols.Retargeting
     /// another FieldSymbol that is responsible for retargeting symbols from one assembly to another. 
     /// It can retarget symbols for multiple assemblies at the same time.
     /// </summary>
-    internal sealed class RetargetingFieldSymbol : FieldSymbol
+    internal sealed class RetargetingFieldSymbol : WrapperFieldSymbol
     {
         /// <summary>
         /// Owning RetargetingModuleSymbol.
         /// </summary>
         private readonly RetargetingModuleSymbol _retargetingModule;
-
-        /// <summary>
-        /// The underlying FieldSymbol, cannot be another RetargetingFieldSymbol.
-        /// </summary>
-        private readonly FieldSymbol _underlyingField;
 
         private ImmutableArray<CustomModifier> _lazyCustomModifiers;
 
@@ -41,13 +36,12 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols.Retargeting
         private DiagnosticInfo _lazyUseSiteDiagnostic = CSDiagnosticInfo.EmptyErrorInfo; // Indicates unknown state. 
 
         public RetargetingFieldSymbol(RetargetingModuleSymbol retargetingModule, FieldSymbol underlyingField)
+            : base (underlyingField)
         {
             Debug.Assert((object)retargetingModule != null);
-            Debug.Assert((object)underlyingField != null);
             Debug.Assert(!(underlyingField is RetargetingFieldSymbol));
 
             _retargetingModule = retargetingModule;
-            _underlyingField = underlyingField;
         }
 
         private RetargetingModuleSymbol.RetargetingSymbolTranslator RetargetingTranslator
@@ -58,25 +52,12 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols.Retargeting
             }
         }
 
-        public FieldSymbol UnderlyingField
-        {
-            get
-            {
-                return _underlyingField;
-            }
-        }
-
         public RetargetingModuleSymbol RetargetingModule
         {
             get
             {
                 return _retargetingModule;
             }
-        }
-
-        public override bool IsImplicitlyDeclared
-        {
-            get { return _underlyingField.IsImplicitlyDeclared; }
         }
 
         internal override TypeSymbol GetFieldType(ConsList<FieldSymbol> fieldsBeingBound)
@@ -97,14 +78,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols.Retargeting
             get
             {
                 return this.RetargetingTranslator.Retarget(_underlyingField.ContainingSymbol);
-            }
-        }
-
-        public override Accessibility DeclaredAccessibility
-        {
-            get
-            {
-                return _underlyingField.DeclaredAccessibility;
             }
         }
 
@@ -134,72 +107,11 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols.Retargeting
             }
         }
 
-        public override string Name
-        {
-            get
-            {
-                return _underlyingField.Name;
-            }
-        }
-
-        internal override bool HasSpecialName
-        {
-            get
-            {
-                return _underlyingField.HasSpecialName;
-            }
-        }
-
-        internal override bool HasRuntimeSpecialName
-        {
-            get
-            {
-                return _underlyingField.HasRuntimeSpecialName;
-            }
-        }
-
-        public override string GetDocumentationCommentXml(CultureInfo preferredCulture = null, bool expandIncludes = false, CancellationToken cancellationToken = default(CancellationToken))
-        {
-            return _underlyingField.GetDocumentationCommentXml(preferredCulture, expandIncludes, cancellationToken);
-        }
-
-        internal override bool IsNotSerialized
-        {
-            get
-            {
-                return _underlyingField.IsNotSerialized;
-            }
-        }
-
-        internal override bool IsMarshalledExplicitly
-        {
-            get
-            {
-                return _underlyingField.IsMarshalledExplicitly;
-            }
-        }
-
         internal override MarshalPseudoCustomAttributeData MarshallingInformation
         {
             get
             {
                 return this.RetargetingTranslator.Retarget(_underlyingField.MarshallingInformation);
-            }
-        }
-
-        internal override ImmutableArray<byte> MarshallingDescriptor
-        {
-            get
-            {
-                return _underlyingField.MarshallingDescriptor;
-            }
-        }
-
-        internal override int? TypeLayoutOffset
-        {
-            get
-            {
-                return _underlyingField.TypeLayoutOffset;
             }
         }
 
@@ -209,75 +121,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols.Retargeting
             {
                 var associated = _underlyingField.AssociatedSymbol;
                 return (object)associated == null ? null : this.RetargetingTranslator.Retarget(associated);
-            }
-        }
-
-        public override bool IsReadOnly
-        {
-            get
-            {
-                return _underlyingField.IsReadOnly;
-            }
-        }
-
-        public override bool IsVolatile
-        {
-            get
-            {
-                return _underlyingField.IsVolatile;
-            }
-        }
-
-        public override bool IsConst
-        {
-            get
-            {
-                return _underlyingField.IsConst;
-            }
-        }
-
-        internal override ObsoleteAttributeData ObsoleteAttributeData
-        {
-            get
-            {
-                return _underlyingField.ObsoleteAttributeData;
-            }
-        }
-
-        public override object ConstantValue
-        {
-            get
-            {
-                return _underlyingField.ConstantValue;
-            }
-        }
-
-        internal override ConstantValue GetConstantValue(ConstantFieldsInProgress inProgress, bool earlyDecodingWellKnownAttributes)
-        {
-            return _underlyingField.GetConstantValue(inProgress, earlyDecodingWellKnownAttributes);
-        }
-
-        public override ImmutableArray<Location> Locations
-        {
-            get
-            {
-                return _underlyingField.Locations;
-            }
-        }
-
-        public override ImmutableArray<SyntaxReference> DeclaringSyntaxReferences
-        {
-            get
-            {
-                return _underlyingField.DeclaringSyntaxReferences;
-            }
-        }
-
-        public override bool IsStatic
-        {
-            get
-            {
-                return _underlyingField.IsStatic;
             }
         }
