+++ src/Compilers/CSharp/Portable/Symbols/Retargeting/RetargetingMethodSymbol.cs
@@ -19,17 +19,12 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols.Retargeting
     /// another MethodSymbol that is responsible for retargeting symbols from one assembly to another. 
     /// It can retarget symbols for multiple assemblies at the same time.
     /// </summary>
-    internal sealed class RetargetingMethodSymbol : MethodSymbol
+    internal sealed class RetargetingMethodSymbol : WrapperMethodSymbol
     {
         /// <summary>
         /// Owning RetargetingModuleSymbol.
         /// </summary>
         private readonly RetargetingModuleSymbol _retargetingModule;
-
-        /// <summary>
-        /// The underlying MethodSymbol, cannot be another RetargetingMethodSymbol.
-        /// </summary>
-        private readonly MethodSymbol _underlyingMethod;
 
         private ImmutableArray<TypeParameterSymbol> _lazyTypeParameters;
 
@@ -53,13 +48,12 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols.Retargeting
         private TypeSymbol _lazyReturnType;
 
         public RetargetingMethodSymbol(RetargetingModuleSymbol retargetingModule, MethodSymbol underlyingMethod)
+            : base(underlyingMethod)
         {
             Debug.Assert((object)retargetingModule != null);
-            Debug.Assert((object)underlyingMethod != null);
             Debug.Assert(!(underlyingMethod is RetargetingMethodSymbol));
 
             _retargetingModule = retargetingModule;
-            _underlyingMethod = underlyingMethod;
         }
 
         private RetargetingModuleSymbol.RetargetingSymbolTranslator RetargetingTranslator
@@ -70,43 +64,11 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols.Retargeting
             }
         }
 
-        public MethodSymbol UnderlyingMethod
-        {
-            get
-            {
-                return _underlyingMethod;
-            }
-        }
-
         public RetargetingModuleSymbol RetargetingModule
         {
             get
             {
                 return _retargetingModule;
-            }
-        }
-
-        public override bool IsVararg
-        {
-            get
-            {
-                return _underlyingMethod.IsVararg;
-            }
-        }
-
-        public override bool IsGenericMethod
-        {
-            get
-            {
-                return _underlyingMethod.IsGenericMethod;
-            }
-        }
-
-        public override int Arity
-        {
-            get
-            {
-                return _underlyingMethod.Arity;
             }
         }
 
@@ -154,14 +116,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols.Retargeting
             }
         }
 
-        internal override RefKind RefKind
-        {
-            get
-            {
-                return _underlyingMethod.RefKind;
-            }
-        }
-
         public override TypeSymbol ReturnType
         {
             get
@@ -183,11 +137,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols.Retargeting
                     _underlyingMethod.ReturnTypeCustomModifiers,
                     ref _lazyCustomModifiers);
             }
-        }
-
-        internal override int ParameterCount
-        {
-            get { return _underlyingMethod.ParameterCount; }
         }
 
         public override ImmutableArray<ParameterSymbol> Parameters
@@ -234,22 +183,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols.Retargeting
             }
         }
 
-        public override bool IsExtensionMethod
-        {
-            get
-            {
-                return _underlyingMethod.IsExtensionMethod;
-            }
-        }
-
-        public override bool HidesBaseMethodsByName
-        {
-            get
-            {
-                return _underlyingMethod.HidesBaseMethodsByName;
-            }
-        }
-
         public override Symbol ContainingSymbol
         {
             get
@@ -258,146 +191,12 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols.Retargeting
             }
         }
 
-        public override ImmutableArray<Location> Locations
-        {
-            get
-            {
-                return _underlyingMethod.Locations;
-            }
-        }
-
-        public override ImmutableArray<SyntaxReference> DeclaringSyntaxReferences
-        {
-            get
-            {
-                return _underlyingMethod.DeclaringSyntaxReferences;
-            }
-        }
-
-        public override Accessibility DeclaredAccessibility
-        {
-            get
-            {
-                return _underlyingMethod.DeclaredAccessibility;
-            }
-        }
-
-        public override bool IsStatic
-        {
-            get
-            {
-                return _underlyingMethod.IsStatic;
-            }
-        }
-
-        public override bool IsVirtual
-        {
-            get
-            {
-                return _underlyingMethod.IsVirtual;
-            }
-        }
-
-        public override bool IsAsync
-        {
-            get
-            {
-                return _underlyingMethod.IsAsync;
-            }
-        }
-
-        public override bool IsOverride
-        {
-            get
-            {
-                return _underlyingMethod.IsOverride;
-            }
-        }
-
-        public override bool IsAbstract
-        {
-            get
-            {
-                return _underlyingMethod.IsAbstract;
-            }
-        }
-
-        public override bool IsSealed
-        {
-            get
-            {
-                return _underlyingMethod.IsSealed;
-            }
-        }
-
-        public override bool IsExtern
-        {
-            get
-            {
-                return _underlyingMethod.IsExtern;
-            }
-        }
-
-        public override bool IsImplicitlyDeclared
-        {
-            get
-            {
-                return _underlyingMethod.IsImplicitlyDeclared;
-            }
-        }
-
-        internal sealed override bool IsMetadataVirtual(bool ignoreInterfaceImplementationChanges = false)
-        {
-            return _underlyingMethod.IsMetadataVirtual(ignoreInterfaceImplementationChanges);
-        }
-
-        internal override bool IsMetadataFinal
-        {
-            get
-            {
-                return _underlyingMethod.IsMetadataFinal;
-            }
-        }
-
-        internal sealed override bool IsMetadataNewSlot(bool ignoreInterfaceImplementationChanges = false)
-        {
-            return _underlyingMethod.IsMetadataNewSlot(ignoreInterfaceImplementationChanges);
-        }
-
-        internal override bool RequiresSecurityObject
-        {
-            get
-            {
-                return _underlyingMethod.RequiresSecurityObject;
-            }
-        }
-
-        public override DllImportData GetDllImportData()
-        {
-            return _underlyingMethod.GetDllImportData();
-        }
-
         internal override MarshalPseudoCustomAttributeData ReturnValueMarshallingInformation
         {
             get
             {
                 return _retargetingModule.RetargetingTranslator.Retarget(_underlyingMethod.ReturnValueMarshallingInformation);
             }
-        }
-
-        internal override bool HasDeclarativeSecurity
-        {
-            get { return _underlyingMethod.HasDeclarativeSecurity; }
-        }
-
-        internal override IEnumerable<Microsoft.Cci.SecurityAttribute> GetSecurityInformation()
-        {
-            return _underlyingMethod.GetSecurityInformation();
-        }
-
-        internal override ImmutableArray<string> GetAppliedConditionalSymbols()
-        {
-            return _underlyingMethod.GetAppliedConditionalSymbols();
         }
 
         public override ImmutableArray<CSharpAttributeData> GetAttributes()
@@ -416,14 +215,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols.Retargeting
             return this.RetargetingTranslator.GetRetargetedAttributes(_underlyingMethod.GetReturnTypeAttributes(), ref _lazyReturnTypeCustomAttributes);
         }
 
-        internal override ObsoleteAttributeData ObsoleteAttributeData
-        {
-            get
-            {
-                return _underlyingMethod.ObsoleteAttributeData;
-            }
-        }
-
         public override AssemblySymbol ContainingAssembly
         {
             get
@@ -437,51 +228,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols.Retargeting
             get
             {
                 return _retargetingModule;
-            }
-        }
-
-        public override string Name
-        {
-            get
-            {
-                return _underlyingMethod.Name;
-            }
-        }
-
-        internal override bool HasSpecialName
-        {
-            get
-            {
-                return _underlyingMethod.HasSpecialName;
-            }
-        }
-
-        public override string GetDocumentationCommentXml(CultureInfo preferredCulture = null, bool expandIncludes = false, CancellationToken cancellationToken = default(CancellationToken))
-        {
-            return _underlyingMethod.GetDocumentationCommentXml(preferredCulture, expandIncludes, cancellationToken);
-        }
-
-        internal override System.Reflection.MethodImplAttributes ImplementationAttributes
-        {
-            get
-            {
-                return _underlyingMethod.ImplementationAttributes;
-            }
-        }
-
-        public override MethodKind MethodKind
-        {
-            get
-            {
-                return _underlyingMethod.MethodKind;
-            }
-        }
-
-        internal override Microsoft.Cci.CallingConvention CallingConvention
-        {
-            get
-            {
-                return _underlyingMethod.CallingConvention;
             }
         }
 
@@ -540,47 +286,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols.Retargeting
             }
 
             return _lazyUseSiteDiagnostic;
-        }
-
-        internal override bool IsAccessCheckedOnOverride
-        {
-            get
-            {
-                return _underlyingMethod.IsAccessCheckedOnOverride;
-            }
-        }
-
-        internal override bool IsExternal
-        {
-            get
-            {
-                return _underlyingMethod.IsExternal;
-            }
-        }
-
-        internal override bool HasRuntimeSpecialName
-        {
-            get
-            {
-                return _underlyingMethod.HasRuntimeSpecialName;
-            }
-        }
-
-
-        internal override bool ReturnValueIsMarshalledExplicitly
-        {
-            get
-            {
-                return _underlyingMethod.ReturnValueIsMarshalledExplicitly;
-            }
-        }
-
-        internal override ImmutableArray<byte> ReturnValueMarshallingDescriptor
-        {
-            get
-            {
-                return _underlyingMethod.ReturnValueMarshallingDescriptor;
-            }
         }
 
         internal sealed override CSharpCompilation DeclaringCompilation // perf, not correctness
