+++ src/Compilers/CSharp/Portable/Symbols/Retargeting/RetargetingNamedTypeSymbol.cs
@@ -20,17 +20,12 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols.Retargeting
     /// another NamedTypeSymbol that is responsible for retargeting referenced symbols from one assembly to another. 
     /// It can retarget symbols for multiple assemblies at the same time.
     /// </summary>
-    internal sealed class RetargetingNamedTypeSymbol : NamedTypeSymbol
+    internal sealed class RetargetingNamedTypeSymbol : WrapperNamedTypeSymbol
     {
         /// <summary>
         /// Owning RetargetingModuleSymbol.
         /// </summary>
         private readonly RetargetingModuleSymbol _retargetingModule;
-
-        /// <summary>
-        /// The underlying NamedTypeSymbol, cannot be another RetargetingNamedTypeSymbol.
-        /// </summary>
-        private readonly NamedTypeSymbol _underlyingType;
 
         private ImmutableArray<TypeParameterSymbol> _lazyTypeParameters;
 
@@ -45,13 +40,12 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols.Retargeting
         private DiagnosticInfo _lazyUseSiteDiagnostic = CSDiagnosticInfo.EmptyErrorInfo; // Indicates unknown state. 
 
         public RetargetingNamedTypeSymbol(RetargetingModuleSymbol retargetingModule, NamedTypeSymbol underlyingType)
+            : base(underlyingType) 
         {
             Debug.Assert((object)retargetingModule != null);
-            Debug.Assert((object)underlyingType != null);
             Debug.Assert(!(underlyingType is RetargetingNamedTypeSymbol));
 
             _retargetingModule = retargetingModule;
-            _underlyingType = underlyingType;
         }
 
         private RetargetingModuleSymbol.RetargetingSymbolTranslator RetargetingTranslator
@@ -59,27 +53,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols.Retargeting
             get
             {
                 return _retargetingModule.RetargetingTranslator;
-            }
-        }
-
-        public NamedTypeSymbol UnderlyingNamedType
-        {
-            get
-            {
-                return _underlyingType;
-            }
-        }
-
-        public override bool IsImplicitlyDeclared
-        {
-            get { return _underlyingType.IsImplicitlyDeclared; }
-        }
-
-        public override int Arity
-        {
-            get
-            {
-                return _underlyingType.Arity;
             }
         }
 
@@ -151,50 +124,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols.Retargeting
                 var underlying = _underlyingType.EnumUnderlyingType;
                 return (object)underlying == null ? null : this.RetargetingTranslator.Retarget(underlying, RetargetOptions.RetargetPrimitiveTypesByTypeCode); // comes from field's signature.
             }
-        }
-
-        public override bool MightContainExtensionMethods
-        {
-            get
-            {
-                return _underlyingType.MightContainExtensionMethods;
-            }
-        }
-
-        public override string Name
-        {
-            get
-            {
-                return _underlyingType.Name;
-            }
-        }
-
-        public override string MetadataName
-        {
-            get
-            {
-                return _underlyingType.MetadataName;
-            }
-        }
-        internal override bool HasSpecialName
-        {
-            get
-            {
-                return _underlyingType.HasSpecialName;
-            }
-        }
-
-        internal override bool MangleName
-        {
-            get
-            {
-                return _underlyingType.MangleName;
-            }
-        }
-
-        public override string GetDocumentationCommentXml(CultureInfo preferredCulture = null, bool expandIncludes = false, CancellationToken cancellationToken = default(CancellationToken))
-        {
-            return _underlyingType.GetDocumentationCommentXml(preferredCulture, expandIncludes, cancellationToken);
         }
 
         public override IEnumerable<string> MemberNames
@@ -299,91 +228,11 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols.Retargeting
             return this.RetargetingTranslator.Retarget(_underlyingType.GetTypeMembers(name, arity));
         }
 
-        public override Accessibility DeclaredAccessibility
-        {
-            get
-            {
-                return _underlyingType.DeclaredAccessibility;
-            }
-        }
-
-        public override TypeKind TypeKind
-        {
-            get
-            {
-                return _underlyingType.TypeKind;
-            }
-        }
-
-        internal override bool IsInterface
-        {
-            get
-            {
-                return _underlyingType.IsInterface;
-            }
-        }
-
         public override Symbol ContainingSymbol
         {
             get
             {
                 return this.RetargetingTranslator.Retarget(_underlyingType.ContainingSymbol);
-            }
-        }
-
-        public override ImmutableArray<Location> Locations
-        {
-            get
-            {
-                return _underlyingType.Locations;
-            }
-        }
-
-        public override ImmutableArray<SyntaxReference> DeclaringSyntaxReferences
-        {
-            get
-            {
-                return _underlyingType.DeclaringSyntaxReferences;
-            }
-        }
-
-        public override bool IsStatic
-        {
-            get
-            {
-                return _underlyingType.IsStatic;
-            }
-        }
-
-        public override bool IsAbstract
-        {
-            get
-            {
-                return _underlyingType.IsAbstract;
-            }
-        }
-
-        internal override bool IsMetadataAbstract
-        {
-            get
-            {
-                return _underlyingType.IsMetadataAbstract;
-            }
-        }
-
-        public override bool IsSealed
-        {
-            get
-            {
-                return _underlyingType.IsSealed;
-            }
-        }
-
-        internal override bool IsMetadataSealed
-        {
-            get
-            {
-                return _underlyingType.IsMetadataSealed;
             }
         }
 
@@ -526,64 +375,9 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols.Retargeting
             get { return _underlyingType.IsComImport; }
         }
 
-        internal override ObsoleteAttributeData ObsoleteAttributeData
-        {
-            get { return _underlyingType.ObsoleteAttributeData; }
-        }
-
-        internal override bool ShouldAddWinRTMembers
-        {
-            get { return _underlyingType.ShouldAddWinRTMembers; }
-        }
-
-        internal override bool IsWindowsRuntimeImport
-        {
-            get { return _underlyingType.IsWindowsRuntimeImport; }
-        }
-
-        internal override TypeLayout Layout
-        {
-            get { return _underlyingType.Layout; }
-        }
-
-        internal override CharSet MarshallingCharSet
-        {
-            get { return _underlyingType.MarshallingCharSet; }
-        }
-
-        internal override bool IsSerializable
-        {
-            get { return _underlyingType.IsSerializable; }
-        }
-
-        internal override bool HasDeclarativeSecurity
-        {
-            get { return _underlyingType.HasDeclarativeSecurity; }
-        }
-
-        internal override IEnumerable<Microsoft.Cci.SecurityAttribute> GetSecurityInformation()
-        {
-            return _underlyingType.GetSecurityInformation();
-        }
-
-        internal override ImmutableArray<string> GetAppliedConditionalSymbols()
-        {
-            return _underlyingType.GetAppliedConditionalSymbols();
-        }
-
-        internal override AttributeUsageInfo GetAttributeUsageInfo()
-        {
-            return _underlyingType.GetAttributeUsageInfo();
-        }
-
         internal sealed override CSharpCompilation DeclaringCompilation // perf, not correctness
         {
             get { return null; }
-        }
-
-        internal override bool GetGuidString(out string guidString)
-        {
-            return _underlyingType.GetGuidString(out guidString);
         }
     }
 }
