+++ src/Compilers/CSharp/Portable/Symbols/Retargeting/RetargetingParameterSymbol.cs
@@ -15,9 +15,8 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols.Retargeting
     /// another ParameterSymbol that is responsible for retargeting symbols from one assembly to another. 
     /// It can retarget symbols for multiple assemblies at the same time.
     /// </summary>
-    internal abstract class RetargetingParameterSymbol : ParameterSymbol
+    internal abstract class RetargetingParameterSymbol : WrapperParameterSymbol
     {
-        private readonly ParameterSymbol _underlyingParameter;
         private ImmutableArray<CustomModifier> _lazyCustomModifiers;
 
         /// <summary>
@@ -26,18 +25,9 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols.Retargeting
         private ImmutableArray<CSharpAttributeData> _lazyCustomAttributes;
 
         protected RetargetingParameterSymbol(ParameterSymbol underlyingParameter)
+            : base(underlyingParameter)
         {
             Debug.Assert(!(underlyingParameter is RetargetingParameterSymbol));
-            _underlyingParameter = underlyingParameter;
-        }
-
-        // test only
-        internal ParameterSymbol UnderlyingParameter
-        {
-            get
-            {
-                return _underlyingParameter;
-            }
         }
 
         protected abstract RetargetingModuleSymbol RetargetingModule
@@ -133,115 +123,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols.Retargeting
         {
             get { return null; }
         }
-
-        #region Forwarded
-
-        internal sealed override ConstantValue ExplicitDefaultConstantValue
-        {
-            get { return _underlyingParameter.ExplicitDefaultConstantValue; }
-        }
-
-        public sealed override RefKind RefKind
-        {
-            get { return _underlyingParameter.RefKind; }
-        }
-
-        internal sealed override bool IsMetadataIn
-        {
-            get { return _underlyingParameter.IsMetadataIn; }
-        }
-
-        internal sealed override bool IsMetadataOut
-        {
-            get { return _underlyingParameter.IsMetadataOut; }
-        }
-
-        public sealed override ImmutableArray<Location> Locations
-        {
-            get { return _underlyingParameter.Locations; }
-        }
-
-        public sealed override ImmutableArray<SyntaxReference> DeclaringSyntaxReferences
-        {
-            get { return _underlyingParameter.DeclaringSyntaxReferences; }
-        }
-
-        internal override void AddSynthesizedAttributes(ModuleCompilationState compilationState, ref ArrayBuilder<SynthesizedAttributeData> attributes)
-        {
-            _underlyingParameter.AddSynthesizedAttributes(compilationState, ref attributes);
-        }
-
-        public override int Ordinal
-        {
-            get { return _underlyingParameter.Ordinal; }
-        }
-
-        public override bool IsParams
-        {
-            get { return _underlyingParameter.IsParams; }
-        }
-
-        internal override bool IsMetadataOptional
-        {
-            get { return _underlyingParameter.IsMetadataOptional; }
-        }
-
-        public override bool IsImplicitlyDeclared
-        {
-            get { return _underlyingParameter.IsImplicitlyDeclared; }
-        }
-
-        public sealed override string Name
-        {
-            get { return _underlyingParameter.Name; }
-        }
-
-        public sealed override string MetadataName
-        {
-            get { return _underlyingParameter.MetadataName; }
-        }
-
-        public override string GetDocumentationCommentXml(CultureInfo preferredCulture = null, bool expandIncludes = false, CancellationToken cancellationToken = default(CancellationToken))
-        {
-            return _underlyingParameter.GetDocumentationCommentXml(preferredCulture, expandIncludes, cancellationToken);
-        }
-
-        internal sealed override UnmanagedType MarshallingType
-        {
-            get { return _underlyingParameter.MarshallingType; }
-        }
-
-        internal sealed override bool IsIDispatchConstant
-        {
-            get { return _underlyingParameter.IsIDispatchConstant; }
-        }
-
-        internal sealed override bool IsIUnknownConstant
-        {
-            get { return _underlyingParameter.IsIUnknownConstant; }
-        }
-
-        internal sealed override bool IsCallerLineNumber
-        {
-            get { return _underlyingParameter.IsCallerLineNumber; }
-        }
-
-        internal sealed override bool IsCallerFilePath
-        {
-            get { return _underlyingParameter.IsCallerFilePath; }
-        }
-
-        internal sealed override bool IsCallerMemberName
-        {
-            get { return _underlyingParameter.IsCallerMemberName; }
-        }
-
-        internal sealed override ushort CountOfCustomModifiersPrecedingByRef
-        {
-            get { return _underlyingParameter.CountOfCustomModifiersPrecedingByRef; }
-        }
-
-        #endregion
     }
 
     internal sealed class RetargetingMethodParameterSymbol : RetargetingParameterSymbol
