+++ src/Compilers/CSharp/Portable/Symbols/SubstitutedFieldSymbol.cs
@@ -12,80 +12,25 @@ using System.Globalization;
 
 namespace Microsoft.CodeAnalysis.CSharp.Symbols
 {
-    internal sealed class SubstitutedFieldSymbol : FieldSymbol
+    internal sealed class SubstitutedFieldSymbol : WrapperFieldSymbol
     {
         private readonly SubstitutedNamedTypeSymbol _containingType;
-        private readonly FieldSymbol _originalDefinition;
-
         private TypeSymbol _lazyType;
 
         internal SubstitutedFieldSymbol(SubstitutedNamedTypeSymbol containingType, FieldSymbol substitutedFrom)
+            : base((FieldSymbol)substitutedFrom.OriginalDefinition)
         {
             _containingType = containingType;
-            _originalDefinition = substitutedFrom.OriginalDefinition as FieldSymbol;
         }
 
         internal override TypeSymbol GetFieldType(ConsList<FieldSymbol> fieldsBeingBound)
         {
             if ((object)_lazyType == null)
             {
-                Interlocked.CompareExchange(ref _lazyType, _containingType.TypeSubstitution.SubstituteType(_originalDefinition.GetFieldType(fieldsBeingBound)).Type, null);
+                Interlocked.CompareExchange(ref _lazyType, _containingType.TypeSubstitution.SubstituteType(OriginalDefinition.GetFieldType(fieldsBeingBound)).Type, null);
             }
 
             return _lazyType;
-        }
-
-        public override string Name
-        {
-            get
-            {
-                return _originalDefinition.Name;
-            }
-        }
-
-        public override string GetDocumentationCommentXml(CultureInfo preferredCulture = null, bool expandIncludes = false, CancellationToken cancellationToken = default(CancellationToken))
-        {
-            return _originalDefinition.GetDocumentationCommentXml(preferredCulture, expandIncludes, cancellationToken);
-        }
-
-        internal override bool HasSpecialName
-        {
-            get
-            {
-                return _originalDefinition.HasSpecialName;
-            }
-        }
-
-        internal override bool HasRuntimeSpecialName
-        {
-            get
-            {
-                return _originalDefinition.HasRuntimeSpecialName;
-            }
-        }
-
-        internal override bool IsNotSerialized
-        {
-            get
-            {
-                return _originalDefinition.IsNotSerialized;
-            }
-        }
-
-        internal override MarshalPseudoCustomAttributeData MarshallingInformation
-        {
-            get
-            {
-                return _originalDefinition.MarshallingInformation;
-            }
-        }
-
-        internal override int? TypeLayoutOffset
-        {
-            get
-            {
-                return _originalDefinition.TypeLayoutOffset;
-            }
         }
 
         public override Symbol ContainingSymbol
@@ -108,36 +53,20 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
         {
             get
             {
-                return _originalDefinition.OriginalDefinition;
-            }
-        }
-
-        public override ImmutableArray<Location> Locations
-        {
-            get
-            {
-                return _originalDefinition.Locations;
-            }
-        }
-
-        public override ImmutableArray<SyntaxReference> DeclaringSyntaxReferences
-        {
-            get
-            {
-                return _originalDefinition.DeclaringSyntaxReferences;
+                return _underlyingField;
             }
         }
 
         public override ImmutableArray<CSharpAttributeData> GetAttributes()
         {
-            return _originalDefinition.GetAttributes();
+            return OriginalDefinition.GetAttributes();
         }
 
         public override Symbol AssociatedSymbol
         {
             get
             {
-                Symbol underlying = _originalDefinition.AssociatedSymbol;
+                Symbol underlying = OriginalDefinition.AssociatedSymbol;
 
                 if ((object)underlying == null)
                 {
@@ -148,80 +77,11 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
             }
         }
 
-        public override bool IsStatic
-        {
-            get
-            {
-                return _originalDefinition.IsStatic;
-            }
-        }
-
-        public override bool IsReadOnly
-        {
-            get
-            {
-                return _originalDefinition.IsReadOnly;
-            }
-        }
-
-        public override bool IsConst
-        {
-            get
-            {
-                return _originalDefinition.IsConst;
-            }
-        }
-
-        internal override ObsoleteAttributeData ObsoleteAttributeData
-        {
-            get
-            {
-                return _originalDefinition.ObsoleteAttributeData;
-            }
-        }
-
-        public override object ConstantValue
-        {
-            get
-            {
-                return _originalDefinition.ConstantValue;
-            }
-        }
-
-        internal override ConstantValue GetConstantValue(ConstantFieldsInProgress inProgress, bool earlyDecodingWellKnownAttributes)
-        {
-            return _originalDefinition.GetConstantValue(inProgress, earlyDecodingWellKnownAttributes);
-        }
-
-        public override bool IsVolatile
-        {
-            get
-            {
-                return _originalDefinition.IsVolatile;
-            }
-        }
-
-        public override bool IsImplicitlyDeclared
-        {
-            get
-            {
-                return _originalDefinition.IsImplicitlyDeclared;
-            }
-        }
-
-        public override Accessibility DeclaredAccessibility
-        {
-            get
-            {
-                return _originalDefinition.DeclaredAccessibility;
-            }
-        }
-
         public override ImmutableArray<CustomModifier> CustomModifiers
         {
             get
             {
-                return _containingType.TypeSubstitution.SubstituteCustomModifiers(_originalDefinition.Type, _originalDefinition.CustomModifiers);
+                return _containingType.TypeSubstitution.SubstituteCustomModifiers(OriginalDefinition.Type, OriginalDefinition.CustomModifiers);
             }
         }
 
@@ -231,7 +91,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
             // containing a fixed-size buffer.  Given the rarity there would be little
             // benefit to "optimizing" the performance of this by caching the
             // translated implementation type.
-            return (NamedTypeSymbol)_containingType.TypeSubstitution.SubstituteType(_originalDefinition.FixedImplementationType(emitModule)).Type;
+            return (NamedTypeSymbol)_containingType.TypeSubstitution.SubstituteType(OriginalDefinition.FixedImplementationType(emitModule)).Type;
         }
 
         public override bool Equals(object obj)
@@ -242,12 +102,12 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
             }
 
             var other = obj as SubstitutedFieldSymbol;
-            return (object)other != null && _containingType == other._containingType && _originalDefinition == other._originalDefinition;
+            return (object)other != null && _containingType == other._containingType && OriginalDefinition == other.OriginalDefinition;
         }
 
         public override int GetHashCode()
         {
-            return Hash.Combine(_containingType, _originalDefinition.GetHashCode());
+            return Hash.Combine(_containingType, OriginalDefinition.GetHashCode());
         }
     }
 }
