+++ src/Compilers/CSharp/Portable/Symbols/SubstitutedMethodSymbol.cs
@@ -1,5 +1,6 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
+using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Diagnostics;
@@ -19,10 +20,9 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
     // C<X> is a ConstructedTypeSymbol.
     // C<X>.M<> is a SubstitutedMethodSymbol. It has parameters of types X and U.
     // C<X>.M<Y> is a ConstructedMethodSymbol.
-    internal class SubstitutedMethodSymbol : MethodSymbol
+    internal class SubstitutedMethodSymbol : WrapperMethodSymbol
     {
         private readonly NamedTypeSymbol _containingType;
-        protected readonly MethodSymbol originalDefinition;
         private readonly TypeMap _inputMap;
         private readonly MethodSymbol _constructedFrom;
 
@@ -43,10 +43,10 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
         }
 
         protected SubstitutedMethodSymbol(NamedTypeSymbol containingSymbol, TypeMap map, MethodSymbol originalDefinition, MethodSymbol constructedFrom)
+            : base (originalDefinition)
         {
             Debug.Assert(originalDefinition.IsDefinition);
             _containingType = containingSymbol;
-            this.originalDefinition = originalDefinition;
             _inputMap = map;
             if ((object)constructedFrom != null)
             {
@@ -67,11 +67,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
             {
                 return _constructedFrom;
             }
-        }
-
-        public override string GetDocumentationCommentXml(CultureInfo preferredCulture = null, bool expandIncludes = false, CancellationToken cancellationToken = default(CancellationToken))
-        {
-            return originalDefinition.GetDocumentationCommentXml(preferredCulture, expandIncludes, cancellationToken);
         }
 
         private TypeMap Map
@@ -103,114 +98,25 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
             Debug.Assert(ReferenceEquals(_constructedFrom, this));
 
             // We're creating a new unconstructed Method from another; alpha-rename type parameters.
-            var newMap = _inputMap.WithAlphaRename(this.originalDefinition, this, out typeParameters);
+            var newMap = _inputMap.WithAlphaRename(this.OriginalDefinition, this, out typeParameters);
 
             var prevMap = Interlocked.CompareExchange(ref _lazyMap, newMap, null);
             if (prevMap != null)
             {
                 // There is a race with another thread who has already set the map
                 // need to ensure that typeParameters, matches the map
-                typeParameters = prevMap.SubstituteTypeParameters(this.originalDefinition.TypeParameters);
+                typeParameters = prevMap.SubstituteTypeParameters(this.OriginalDefinition.TypeParameters);
             }
 
             ImmutableInterlocked.InterlockedCompareExchange(ref _lazyTypeParameters, typeParameters, default(ImmutableArray<TypeParameterSymbol>));
             Debug.Assert(_lazyTypeParameters != null);
         }
 
-        internal sealed override Microsoft.Cci.CallingConvention CallingConvention
-        {
-            get
-            {
-                return originalDefinition.CallingConvention;
-            }
-        }
-
-        public sealed override int Arity
-        {
-            get
-            {
-                return originalDefinition.Arity;
-            }
-        }
-
-        public sealed override string Name
-        {
-            get
-            {
-                return originalDefinition.Name;
-            }
-        }
-
-        internal sealed override bool HasSpecialName
-        {
-            get
-            {
-                return originalDefinition.HasSpecialName;
-            }
-        }
-
-        internal sealed override System.Reflection.MethodImplAttributes ImplementationAttributes
-        {
-            get
-            {
-                return originalDefinition.ImplementationAttributes;
-            }
-        }
-
-        internal sealed override bool RequiresSecurityObject
-        {
-            get
-            {
-                return originalDefinition.RequiresSecurityObject;
-            }
-        }
-
-        public sealed override DllImportData GetDllImportData()
-        {
-            return originalDefinition.GetDllImportData();
-        }
-
-        internal sealed override MarshalPseudoCustomAttributeData ReturnValueMarshallingInformation
-        {
-            get { return originalDefinition.ReturnValueMarshallingInformation; }
-        }
-
-        internal sealed override bool HasDeclarativeSecurity
-        {
-            get { return originalDefinition.HasDeclarativeSecurity; }
-        }
-
-        internal sealed override IEnumerable<Microsoft.Cci.SecurityAttribute> GetSecurityInformation()
-        {
-            return originalDefinition.GetSecurityInformation();
-        }
-
-        internal sealed override ImmutableArray<string> GetAppliedConditionalSymbols()
-        {
-            return originalDefinition.GetAppliedConditionalSymbols();
-        }
-
         public sealed override AssemblySymbol ContainingAssembly
         {
             get
             {
-                return originalDefinition.ContainingAssembly;
-            }
-        }
-
-        public sealed override ImmutableArray<Location> Locations
-        {
-            get
-            {
-                return originalDefinition.Locations;
-            }
-        }
-
-        public sealed override ImmutableArray<SyntaxReference> DeclaringSyntaxReferences
-        {
-            get
-            {
-                return originalDefinition.DeclaringSyntaxReferences;
+                return OriginalDefinition.ContainingAssembly;
             }
         }
 
@@ -226,79 +132,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
         {
             get
             {
-                return originalDefinition;
-            }
-        }
-
-        public sealed override bool IsExtern
-        {
-            get
-            {
-                return originalDefinition.IsExtern;
-            }
-        }
-
-        public sealed override bool IsSealed
-        {
-            get
-            {
-                return originalDefinition.IsSealed;
-            }
-        }
-
-        public sealed override bool IsVirtual
-        {
-            get
-            {
-                return originalDefinition.IsVirtual;
-            }
-        }
-
-        public sealed override bool IsAsync
-        {
-            get
-            {
-                return originalDefinition.IsAsync;
-            }
-        }
-
-        public sealed override bool IsAbstract
-        {
-            get
-            {
-                return originalDefinition.IsAbstract;
-            }
-        }
-
-        public sealed override bool IsOverride
-        {
-            get
-            {
-                return originalDefinition.IsOverride;
-            }
-        }
-
-        public sealed override bool IsStatic
-        {
-            get
-            {
-                return originalDefinition.IsStatic;
-            }
-        }
-
-        public sealed override bool IsExtensionMethod
-        {
-            get
-            {
-                return originalDefinition.IsExtensionMethod;
-            }
-        }
-
-        internal override ObsoleteAttributeData ObsoleteAttributeData
-        {
-            get
-            {
-                return originalDefinition.ObsoleteAttributeData;
+                return _underlyingMethod;
             }
         }
 
@@ -306,7 +140,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
         {
             get
             {
-                var method = originalDefinition.ReducedFrom;
+                var method = OriginalDefinition.ReducedFrom;
                 return ((object)method == null) ? null : method.Construct(this.TypeArguments);
             }
         }
@@ -328,9 +162,9 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
         public override TypeSymbol GetTypeInferredDuringReduction(TypeParameterSymbol reducedFromTypeParameter)
         {
             // This will throw if API shouldn't be supported or there is a problem with the argument.
-            var notUsed = originalDefinition.GetTypeInferredDuringReduction(reducedFromTypeParameter);
+            var notUsed = OriginalDefinition.GetTypeInferredDuringReduction(reducedFromTypeParameter);
 
-            Debug.Assert((object)notUsed == null && (object)originalDefinition.ReducedFrom != null);
+            Debug.Assert((object)notUsed == null && (object)OriginalDefinition.ReducedFrom != null);
             return this.TypeArguments[reducedFromTypeParameter.Ordinal];
         }
 
@@ -338,42 +172,8 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
         {
             get
             {
-                return originalDefinition.ReducedFrom;
+                return OriginalDefinition.ReducedFrom;
             }
-        }
-
-        public sealed override bool HidesBaseMethodsByName
-        {
-            get
-            {
-                return originalDefinition.HidesBaseMethodsByName;
-            }
-        }
-
-        public sealed override Accessibility DeclaredAccessibility
-        {
-            get
-            {
-                return originalDefinition.DeclaredAccessibility;
-            }
-        }
-
-        internal sealed override bool IsMetadataVirtual(bool ignoreInterfaceImplementationChanges = false)
-        {
-            return originalDefinition.IsMetadataVirtual(ignoreInterfaceImplementationChanges);
-        }
-
-        internal override bool IsMetadataFinal
-        {
-            get
-            {
-                return originalDefinition.IsMetadataFinal;
-            }
-        }
-
-        internal sealed override bool IsMetadataNewSlot(bool ignoreInterfaceImplementationChanges = false)
-        {
-            return originalDefinition.IsMetadataNewSlot(ignoreInterfaceImplementationChanges);
         }
 
         public sealed override Symbol ContainingSymbol
@@ -394,23 +194,20 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
 
         public sealed override ImmutableArray<CSharpAttributeData> GetAttributes()
         {
-            return this.originalDefinition.GetAttributes();
+            return this.OriginalDefinition.GetAttributes();
+        }
+
+        public override ImmutableArray<CSharpAttributeData> GetReturnTypeAttributes()
+        {
+            return this.OriginalDefinition.GetReturnTypeAttributes();
         }
 
         public sealed override Symbol AssociatedSymbol
         {
             get
             {
-                Symbol underlying = originalDefinition.AssociatedSymbol;
+                Symbol underlying = OriginalDefinition.AssociatedSymbol;
                 return ((object)underlying == null) ? null : underlying.SymbolAsMember(ContainingType);
-            }
-        }
-
-        public sealed override MethodKind MethodKind
-        {
-            get
-            {
-                return originalDefinition.MethodKind;
             }
         }
 
@@ -418,47 +215,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
         {
             get
             {
-                return originalDefinition.ReturnsVoid;
-            }
-        }
-
-        public sealed override bool IsGenericMethod
-        {
-            get
-            {
-                return originalDefinition.IsGenericMethod;
-            }
-        }
-
-        public sealed override bool IsImplicitlyDeclared
-        {
-            get
-            {
-                return this.originalDefinition.IsImplicitlyDeclared;
-            }
-        }
-
-        internal sealed override bool GenerateDebugInfo
-        {
-            get
-            {
-                return this.originalDefinition.GenerateDebugInfo;
-            }
-        }
-
-        public sealed override bool IsVararg
-        {
-            get
-            {
-                return originalDefinition.IsVararg;
-            }
-        }
-
-        internal sealed override RefKind RefKind
-        {
-            get
-            {
-                return originalDefinition.RefKind;
+                return OriginalDefinition.ReturnsVoid;
             }
         }
 
@@ -472,7 +229,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
                     return returnType;
                 }
 
-                returnType = Map.SubstituteType(originalDefinition.ReturnType).Type;
+                returnType = Map.SubstituteType(OriginalDefinition.ReturnType).Type;
                 return Interlocked.CompareExchange(ref _lazyReturnType, returnType, null) ?? returnType;
             }
         }
@@ -481,13 +238,8 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
         {
             get
             {
-                return Map.SubstituteCustomModifiers(originalDefinition.ReturnType, originalDefinition.ReturnTypeCustomModifiers);
+                return Map.SubstituteCustomModifiers(OriginalDefinition.ReturnType, OriginalDefinition.ReturnTypeCustomModifiers);
             }
-        }
-
-        internal sealed override int ParameterCount
-        {
-            get { return this.originalDefinition.ParameterCount; }
         }
 
         public sealed override ImmutableArray<ParameterSymbol> Parameters
@@ -505,7 +257,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
 
         internal sealed override bool IsExplicitInterfaceImplementation
         {
-            get { return this.originalDefinition.IsExplicitInterfaceImplementation; }
+            get { return this.OriginalDefinition.IsExplicitInterfaceImplementation; }
         }
 
         public sealed override ImmutableArray<MethodSymbol> ExplicitInterfaceImplementations
@@ -521,7 +273,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
                 {
                     ImmutableInterlocked.InterlockedCompareExchange(
                         ref _lazyExplicitInterfaceImplementations,
-                        ExplicitInterfaceHelpers.SubstituteExplicitInterfaceImplementations(this.originalDefinition.ExplicitInterfaceImplementations, Map),
+                        ExplicitInterfaceHelpers.SubstituteExplicitInterfaceImplementations(this.OriginalDefinition.ExplicitInterfaceImplementations, Map),
                         default(ImmutableArray<MethodSymbol>));
                 }
                 return _lazyExplicitInterfaceImplementations;
@@ -545,7 +297,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
 
         internal sealed override bool CallsAreOmitted(SyntaxTree syntaxTree)
         {
-            return originalDefinition.CallsAreOmitted(syntaxTree);
+            return OriginalDefinition.CallsAreOmitted(syntaxTree);
         }
 
         internal sealed override TypeMap TypeSubstitution
@@ -561,7 +313,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
             // There should never be any reason to call this in normal compilation
             // scenarios, but the behavior should be sensible if it does occur.
             ParameterSymbol originalThisParameter;
-            if (!originalDefinition.TryGetThisParameter(out originalThisParameter))
+            if (!OriginalDefinition.TryGetThisParameter(out originalThisParameter))
             {
                 thisParameter = null;
                 return false;
@@ -575,7 +327,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
 
         private ImmutableArray<ParameterSymbol> SubstituteParameters()
         {
-            var unsubstitutedParameters = originalDefinition.Parameters;
+            var unsubstitutedParameters = OriginalDefinition.Parameters;
             int count = unsubstitutedParameters.Length;
 
             if (count == 0)
