+++ src/Compilers/CSharp/Portable/Symbols/SubstitutedPropertySymbol.cs
@@ -9,26 +9,17 @@ using Microsoft.CodeAnalysis.Text;
 
 namespace Microsoft.CodeAnalysis.CSharp.Symbols
 {
-    internal sealed class SubstitutedPropertySymbol : PropertySymbol
+    internal sealed class SubstitutedPropertySymbol : WrapperPropertySymbol
     {
-        private readonly PropertySymbol _originalDefinition;
         private readonly SubstitutedNamedTypeSymbol _containingType;
 
         private TypeSymbol _lazyType;
         private ImmutableArray<ParameterSymbol> _lazyParameters;
 
         internal SubstitutedPropertySymbol(SubstitutedNamedTypeSymbol containingType, PropertySymbol originalDefinition)
+            : base(originalDefinition)
         {
             _containingType = containingType;
-            _originalDefinition = originalDefinition;
-        }
-
-        internal override RefKind RefKind
-        {
-            get
-            {
-                return _originalDefinition.RefKind;
-            }
         }
 
         public override TypeSymbol Type
@@ -37,32 +28,11 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
             {
                 if ((object)_lazyType == null)
                 {
-                    Interlocked.CompareExchange(ref _lazyType, _containingType.TypeSubstitution.SubstituteType(_originalDefinition.Type).Type, null);
+                    Interlocked.CompareExchange(ref _lazyType, _containingType.TypeSubstitution.SubstituteType(OriginalDefinition.Type).Type, null);
                 }
 
                 return _lazyType;
             }
-        }
-
-        public override string Name
-        {
-            get
-            {
-                return _originalDefinition.Name;
-            }
-        }
-
-        internal override bool HasSpecialName
-        {
-            get
-            {
-                return _originalDefinition.HasSpecialName;
-            }
-        }
-
-        public override string GetDocumentationCommentXml(CultureInfo preferredCulture = null, bool expandIncludes = false, CancellationToken cancellationToken = default(CancellationToken))
-        {
-            return _originalDefinition.GetDocumentationCommentXml(preferredCulture, expandIncludes, cancellationToken);
         }
 
         public override Symbol ContainingSymbol
@@ -85,82 +55,18 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
         {
             get
             {
-                return _originalDefinition;
-            }
-        }
-
-        public override ImmutableArray<Location> Locations
-        {
-            get
-            {
-                return _originalDefinition.Locations;
-            }
-        }
-
-        public override ImmutableArray<SyntaxReference> DeclaringSyntaxReferences
-        {
-            get
-            {
-                return _originalDefinition.DeclaringSyntaxReferences;
+                return _underlyingProperty;
             }
         }
 
         public override ImmutableArray<CSharpAttributeData> GetAttributes()
         {
-            return _originalDefinition.GetAttributes();
-        }
-
-        public override bool IsStatic
-        {
-            get
-            {
-                return _originalDefinition.IsStatic;
-            }
-        }
-
-        public override bool IsExtern
-        {
-            get { return _originalDefinition.IsExtern; }
-        }
-
-        public override bool IsSealed
-        {
-            get { return _originalDefinition.IsSealed; }
-        }
-
-        public override bool IsAbstract
-        {
-            get { return _originalDefinition.IsAbstract; }
-        }
-
-        public override bool IsVirtual
-        {
-            get { return _originalDefinition.IsVirtual; }
-        }
-
-        public override bool IsOverride
-        {
-            get { return _originalDefinition.IsOverride; }
-        }
-
-        public override bool IsImplicitlyDeclared
-        {
-            get { return _originalDefinition.IsImplicitlyDeclared; }
-        }
-
-        internal override ObsoleteAttributeData ObsoleteAttributeData
-        {
-            get { return _originalDefinition.ObsoleteAttributeData; }
-        }
-
-        public override bool IsIndexer
-        {
-            get { return _originalDefinition.IsIndexer; }
+            return OriginalDefinition.GetAttributes();
         }
 
         public override ImmutableArray<CustomModifier> TypeCustomModifiers
         {
-            get { return _containingType.TypeSubstitution.SubstituteCustomModifiers(_originalDefinition.Type, _originalDefinition.TypeCustomModifiers); }
+            get { return _containingType.TypeSubstitution.SubstituteCustomModifiers(OriginalDefinition.Type, OriginalDefinition.TypeCustomModifiers); }
         }
 
         public override ImmutableArray<ParameterSymbol> Parameters
@@ -176,12 +82,11 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
             }
         }
 
-
         public override MethodSymbol GetMethod
         {
             get
             {
-                MethodSymbol originalGetMethod = _originalDefinition.GetMethod;
+                MethodSymbol originalGetMethod = OriginalDefinition.GetMethod;
                 return (object)originalGetMethod == null ? null : originalGetMethod.AsMember(_containingType);
             }
         }
@@ -190,14 +95,14 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
         {
             get
             {
-                MethodSymbol originalSetMethod = _originalDefinition.SetMethod;
+                MethodSymbol originalSetMethod = OriginalDefinition.SetMethod;
                 return (object)originalSetMethod == null ? null : originalSetMethod.AsMember(_containingType);
             }
         }
 
         internal override bool IsExplicitInterfaceImplementation
         {
-            get { return _originalDefinition.IsExplicitInterfaceImplementation; }
+            get { return OriginalDefinition.IsExplicitInterfaceImplementation; }
         }
 
         //we want to compute this lazily since it may be expensive for the underlying symbol
@@ -213,29 +118,16 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
                 {
                     ImmutableInterlocked.InterlockedCompareExchange(
                         ref _lazyExplicitInterfaceImplementations,
-                        ExplicitInterfaceHelpers.SubstituteExplicitInterfaceImplementations(_originalDefinition.ExplicitInterfaceImplementations, _containingType.TypeSubstitution),
+                        ExplicitInterfaceHelpers.SubstituteExplicitInterfaceImplementations(OriginalDefinition.ExplicitInterfaceImplementations, _containingType.TypeSubstitution),
                         default(ImmutableArray<PropertySymbol>));
                 }
                 return _lazyExplicitInterfaceImplementations;
             }
         }
 
-        internal override Microsoft.Cci.CallingConvention CallingConvention
-        {
-            get { return _originalDefinition.CallingConvention; }
-        }
-
         internal override bool MustCallMethodsDirectly
         {
-            get { return _originalDefinition.MustCallMethodsDirectly; }
-        }
-
-        public override Accessibility DeclaredAccessibility
-        {
-            get
-            {
-                return _originalDefinition.DeclaredAccessibility;
-            }
+            get { return OriginalDefinition.MustCallMethodsDirectly; }
         }
 
         internal override OverriddenOrHiddenMembersResult OverriddenOrHiddenMembers
@@ -246,13 +138,14 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
                 {
                     Interlocked.CompareExchange(ref _lazyOverriddenOrHiddenMembers, this.MakeOverriddenOrHiddenMembers(), null);
                 }
+
                 return _lazyOverriddenOrHiddenMembers;
             }
         }
 
         private ImmutableArray<ParameterSymbol> SubstituteParameters()
         {
-            var unsubstitutedParameters = _originalDefinition.Parameters;
+            var unsubstitutedParameters = OriginalDefinition.Parameters;
 
             if (unsubstitutedParameters.IsEmpty)
             {
@@ -267,20 +160,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
                     substituted[i] = new SubstitutedParameterSymbol(this, _containingType.TypeSubstitution, unsubstitutedParameters[i]);
                 }
                 return substituted.AsImmutableOrNull();
-            }
-        }
-
-        public override string MetadataName
-        {
-            // We'll never emit this symbol, so it doesn't really
-            // make sense for it to have a metadata name.  However, all
-            // symbols have an implementation of MetadataName (since it
-            // is virtual on Symbol) so we might as well define it in a
-            // consistent way.
-
-            get
-            {
-                return _originalDefinition.MetadataName;
             }
         }
     }
