+++ src/Compilers/CSharp/Portable/Symbols/SubstitutedTypeParameterSymbol.cs
@@ -10,11 +10,10 @@ using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.CSharp.Symbols
 {
-    internal class SubstitutedTypeParameterSymbol : TypeParameterSymbol
+    internal class SubstitutedTypeParameterSymbol : WrapperTypeParameterSymbol
     {
         private readonly Symbol _container;
         private readonly TypeMap _map;
-        private readonly TypeParameterSymbol _substitutedFrom;
         private readonly int _ordinal;
 
 #if DEBUG_ALPHA
@@ -23,24 +22,16 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
 #endif
 
         internal SubstitutedTypeParameterSymbol(Symbol newContainer, TypeMap map, TypeParameterSymbol substitutedFrom, int ordinal)
+            : base(substitutedFrom) 
         {
             _container = newContainer;
             // it is important that we don't use the map here in the constructor, as the map is still being filled
             // in by TypeMap.WithAlphaRename.  Instead, we can use the map lazily when yielding the constraints.
             _map = map;
-            _substitutedFrom = substitutedFrom;
             _ordinal = ordinal;
 #if DEBUG_ALPHA
             _mySequence = _nextSequence++;
 #endif
-        }
-
-        public override TypeParameterKind TypeParameterKind
-        {
-            get
-            {
-                return _substitutedFrom.TypeParameterKind;
-            }
         }
 
         public override Symbol ContainingSymbol
@@ -51,22 +42,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
             }
         }
 
-        public override ImmutableArray<Location> Locations
-        {
-            get
-            {
-                return _substitutedFrom.Locations;
-            }
-        }
-
-        public override ImmutableArray<SyntaxReference> DeclaringSyntaxReferences
-        {
-            get
-            {
-                return _substitutedFrom.DeclaringSyntaxReferences;
-            }
-        }
-
         public override TypeParameterSymbol OriginalDefinition
         {
             get
@@ -74,8 +49,8 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
                 // A substituted type parameter symbol is used as a type parameter of a frame type for lambda-captured
                 // variables within a generic method.  In that case the frame's own type parameter is an original.
                 return
-                    ContainingSymbol.OriginalDefinition != _substitutedFrom.ContainingSymbol.OriginalDefinition ? this :
-                    _substitutedFrom.OriginalDefinition;
+                    ContainingSymbol.OriginalDefinition != _underlyingTypeParameter.ContainingSymbol.OriginalDefinition ? this :
+                    _underlyingTypeParameter.OriginalDefinition;
             }
         }
 
@@ -97,14 +72,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
             }
         }
 
-        public override bool HasConstructorConstraint
-        {
-            get
-            {
-                return _substitutedFrom.HasConstructorConstraint;
-            }
-        }
-
         public override int Ordinal
         {
             get
@@ -113,35 +80,11 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
             }
         }
 
-        public override VarianceKind Variance
-        {
-            get
-            {
-                return _substitutedFrom.Variance;
-            }
-        }
-
-        public override bool HasValueTypeConstraint
-        {
-            get
-            {
-                return _substitutedFrom.HasValueTypeConstraint;
-            }
-        }
-
-        public override bool HasReferenceTypeConstraint
-        {
-            get
-            {
-                return _substitutedFrom.HasReferenceTypeConstraint;
-            }
-        }
-
         public override string Name
         {
             get
             {
-                return _substitutedFrom.Name
+                return base.Name
 #if DEBUG_ALPHA
                     + "#" + _mySequence
 #endif
@@ -149,47 +92,29 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
             }
         }
 
-        public override bool IsImplicitlyDeclared
-        {
-            get
-            {
-                return _substitutedFrom.IsImplicitlyDeclared;
-            }
-        }
-
-        public override string GetDocumentationCommentXml(CultureInfo preferredCulture = null, bool expandIncludes = false, CancellationToken cancellationToken = default(CancellationToken))
-        {
-            return _substitutedFrom.GetDocumentationCommentXml(preferredCulture, expandIncludes, cancellationToken);
-        }
-
         public override ImmutableArray<CSharpAttributeData> GetAttributes()
         {
-            return _substitutedFrom.GetAttributes();
-        }
-
-        internal override void EnsureAllConstraintsAreResolved()
-        {
-            _substitutedFrom.EnsureAllConstraintsAreResolved();
+            return _underlyingTypeParameter.GetAttributes();
         }
 
         internal override ImmutableArray<TypeSymbol> GetConstraintTypes(ConsList<TypeParameterSymbol> inProgress)
         {
-            return _map.SubstituteTypesWithoutModifiers(_substitutedFrom.GetConstraintTypes(inProgress)).WhereAsArray(s_isNotObjectFunc).Distinct();
+            return _map.SubstituteTypesWithoutModifiers(_underlyingTypeParameter.GetConstraintTypes(inProgress)).WhereAsArray(s_isNotObjectFunc).Distinct();
         }
 
         internal override ImmutableArray<NamedTypeSymbol> GetInterfaces(ConsList<TypeParameterSymbol> inProgress)
         {
-            return _map.SubstituteNamedTypes(_substitutedFrom.GetInterfaces(inProgress));
+            return _map.SubstituteNamedTypes(_underlyingTypeParameter.GetInterfaces(inProgress));
         }
 
         internal override NamedTypeSymbol GetEffectiveBaseClass(ConsList<TypeParameterSymbol> inProgress)
         {
-            return _map.SubstituteNamedType(_substitutedFrom.GetEffectiveBaseClass(inProgress));
+            return _map.SubstituteNamedType(_underlyingTypeParameter.GetEffectiveBaseClass(inProgress));
         }
 
         internal override TypeSymbol GetDeducedBaseType(ConsList<TypeParameterSymbol> inProgress)
         {
-            return _map.SubstituteType(_substitutedFrom.GetDeducedBaseType(inProgress)).AsTypeSymbolOnly();
+            return _map.SubstituteType(_underlyingTypeParameter.GetDeducedBaseType(inProgress)).AsTypeSymbolOnly();
         }
 
         private static readonly Func<TypeSymbol, bool> s_isNotObjectFunc = type => type.SpecialType != SpecialType.System_Object;
