+++ src/Compilers/CSharp/Portable/Symbols/Wrappers/WrapperPropertySymbol.cs
@@ -1,0 +1,232 @@
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
+using System.Collections.Generic;
+using System.Collections.Immutable;
+using System.Diagnostics;
+using System.Globalization;
+using System.Threading;
+
+namespace Microsoft.CodeAnalysis.CSharp.Symbols
+{
+    /// <summary>
+    /// Represents a property that is based on another property.
+    /// When inheriting from this class, one shouldn't assume that 
+    /// the default behavior it has is appropriate for every case.
+    /// That behavior should be carefully reviewed and derived type
+    /// should override behavior as appropriate.
+    /// </summary>
+    internal abstract class WrapperPropertySymbol : PropertySymbol
+    {
+        /// <summary>
+        /// The underlying PropertySymbol.
+        /// </summary>
+        protected readonly PropertySymbol _underlyingProperty;
+
+        public WrapperPropertySymbol(PropertySymbol underlyingProperty)
+        {
+            Debug.Assert((object)underlyingProperty != null);
+            _underlyingProperty = underlyingProperty;
+        }
+
+        public PropertySymbol UnderlyingProperty
+        {
+            get
+            {
+                return _underlyingProperty;
+            }
+        }
+
+        public override bool IsImplicitlyDeclared
+        {
+            get { return _underlyingProperty.IsImplicitlyDeclared; }
+        }
+
+        internal override RefKind RefKind
+        {
+            get
+            {
+                return _underlyingProperty.RefKind;
+            }
+        }
+
+        public override abstract TypeSymbol Type
+        {
+            get;
+        }
+
+        public override abstract ImmutableArray<CustomModifier> TypeCustomModifiers
+        {
+            get;
+        }
+
+        public override abstract ImmutableArray<ParameterSymbol> Parameters
+        {
+            get;
+        }
+
+        public override bool IsIndexer
+        {
+            get
+            {
+                return _underlyingProperty.IsIndexer;
+            }
+        }
+
+        public override abstract MethodSymbol GetMethod
+        {
+            get;
+        }
+
+        public override abstract MethodSymbol SetMethod
+        {
+            get;
+        }
+
+        internal override Microsoft.Cci.CallingConvention CallingConvention
+        {
+            get
+            {
+                return _underlyingProperty.CallingConvention;
+            }
+        }
+
+        internal override abstract bool IsExplicitInterfaceImplementation
+        {
+            get;
+        }
+
+        public override abstract ImmutableArray<PropertySymbol> ExplicitInterfaceImplementations
+        {
+            get;
+        }
+
+        public override abstract Symbol ContainingSymbol
+        {
+            get;
+        }
+
+        public override string Name
+        {
+            get
+            {
+                return _underlyingProperty.Name;
+            }
+        }
+
+        internal override bool HasSpecialName
+        {
+            get
+            {
+                return _underlyingProperty.HasSpecialName;
+            }
+        }
+
+        public override string GetDocumentationCommentXml(CultureInfo preferredCulture = null, bool expandIncludes = false, CancellationToken cancellationToken = default(CancellationToken))
+        {
+            return _underlyingProperty.GetDocumentationCommentXml(preferredCulture, expandIncludes, cancellationToken);
+        }
+
+        public override ImmutableArray<Location> Locations
+        {
+            get
+            {
+                return _underlyingProperty.Locations;
+            }
+        }
+
+        public override ImmutableArray<SyntaxReference> DeclaringSyntaxReferences
+        {
+            get
+            {
+                return _underlyingProperty.DeclaringSyntaxReferences;
+            }
+        }
+
+        public override Accessibility DeclaredAccessibility
+        {
+            get
+            {
+                return _underlyingProperty.DeclaredAccessibility;
+            }
+        }
+
+        public override bool IsStatic
+        {
+            get
+            {
+                return _underlyingProperty.IsStatic;
+            }
+        }
+
+        public override bool IsVirtual
+        {
+            get
+            {
+                return _underlyingProperty.IsVirtual;
+            }
+        }
+
+        public override bool IsOverride
+        {
+            get
+            {
+                return _underlyingProperty.IsOverride;
+            }
+        }
+
+        public override bool IsAbstract
+        {
+            get
+            {
+                return _underlyingProperty.IsAbstract;
+            }
+        }
+
+        public override bool IsSealed
+        {
+            get
+            {
+                return _underlyingProperty.IsSealed;
+            }
+        }
+
+        public override bool IsExtern
+        {
+            get
+            {
+                return _underlyingProperty.IsExtern;
+            }
+        }
+
+        internal override ObsoleteAttributeData ObsoleteAttributeData
+        {
+            get
+            {
+                return _underlyingProperty.ObsoleteAttributeData;
+            }
+        }
+
+        public override abstract ImmutableArray<CSharpAttributeData> GetAttributes();
+
+        internal override abstract bool MustCallMethodsDirectly
+        {
+            get;
+        }
+
+        public override string MetadataName
+        {
+            get
+            {
+                return _underlyingProperty.MetadataName;
+            }
+        }
+
+        internal override bool HasRuntimeSpecialName
+        {
+            get
+            {
+                return _underlyingProperty.HasRuntimeSpecialName;
+            }
+        }
+    }
+}
